backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
pytest
pyTelegramBotAPI
python-dotenv
beautifulsoup4
